identify silences (inefficiently)
This seems like it might be a much better
first-pass "note split detector". It
identifies the most important splits with
likely very high accuracy.

There are other important splits that it
likely _won't_ catch (switching frequencies
while keeping amplitude) but those are
actually possible to manually avoid while
recording.

This is a naïvely coded proof-of-concept,
it's currently needlessly expensive.