trying out another noise-cleaning approach
This does _very_ well on melodies that have
clear separation between notes, i.e. no
"sliding".

(For instance, it gets near-perfect results
when whistling the Final Fantasy victory
tune.)

I think it might be better to build on this,
and expand it to detect the somewhat rarer
case of notes changing frequency without
even a brief period of silence.